+++ src/Metadata/Metadata.php
@@ -2,7 +2,10 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilationSource;
+namespace webignition\BasilCompilationSource\Metadata;
+
+use webignition\BasilCompilationSource\ClassDependencyCollection;
+use webignition\BasilCompilationSource\VariablePlaceholderCollection;
 
 class Metadata implements MetadataInterface
 {
