+++ src/SourceInterface.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace webignition\BasilCompilationSource;
 
+use webignition\BasilCompilationSource\Metadata\MetadataInterface;
+
 interface SourceInterface
 {
     public function getMetadata(): MetadataInterface;
